refactor: move ResourceIcon to common/
