+++ packages/frontend/src/components/common/ResourceIcon/index.tsx
